Move properties handling out of serializer
