+++ packages/sdk/client/src/echo/index.ts
@@ -64,3 +64,4 @@ export {
   OBJECT_ID_LENGTH,
   SPACE_ID_LENGTH,
 } from './util';
+export * from './import';
